Fix missing dependencies during setup
I tried to build Pyto in a new Mac, many dependencies weren't installed so now the scripts should take care of that. Compilation of SciPy is not yet possible on Apple Silicon because Docker isn't available and that's required for building Fortran code. Everything else should just work if CocoaPods and Homebrew are installed by running `arch -x86_64 ./setup.sh`.